fix(CartProduct): 장바구니 아이템을 삭제할 때 position이 아니라 id로 삭제하도록 변경

+++ app/src/main/java/woowacourse/shopping/data/DefaultCartRepository.kt
@@ -18,7 +18,7 @@ class DefaultCartRepository(private val dao: CartProductDao) : CartRepository {
         return dao.getAll().map(CartProductEntity::toData)
     }
 
-    override suspend fun deleteCartProduct(id: Int) {
-        dao.delete(id.toLong())
+    override suspend fun deleteCartProduct(id: Long) {
+        dao.delete(id)
     }
 }
