+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModel.kt
@@ -27,7 +27,7 @@ class CartViewModel(
         }
     }
 
-    fun deleteCartProduct(id: Int) {
+    fun deleteCartProduct(id: Long) {
         viewModelScope.launch {
             cartRepository.deleteCartProduct(id)
             _onCartProductDeleted.value = true
